Refactor library to use es2015 plus some extra new syntaxes supported by babel.
- As part of this FB is now an instance of an internal Facebook class.
- Add some more tests.
- Timeout for tests has been increased to 10000ms due to babel parser slowing down mocha-eslint.
