+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules/
 npm-debug.log
+/lib/
 
 .DS_Store
 Thumbs.db
