+++ .npmignore
@@ -1,3 +1,5 @@
+src/
+
 Thumbs.db
 # WebStrom IDE ignore
 .idea/
