+++ package.json
@@ -20,15 +20,23 @@
   "main": "./fb.js",
   "scripts": {
     "lint": "eslint .",
-    "test": "node ./node_modules/mocha/bin/mocha --recursive --reporter spec",
-    "testw": "node ./node_modules/mocha/bin/mocha --recursive --watch --reporter spec"
+    "build": "babel src/ -d lib/",
+    "buildw": "babel -w src/ -d lib/",
+    "test": "npm run build && node ./node_modules/mocha/bin/mocha --recursive ",
+    "prepublish": "npm run build"
   },
   "dependencies": {
+    "babel-runtime": "^6.3.19",
     "debug": "^2.2.0",
     "request": "^2.67.0"
   },
   "devDependencies": {
     "babel-cli": "^6.4.5",
+    "babel-eslint": "^5.0.0-beta6",
+    "babel-plugin-transform-runtime": "^6.4.3",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-stage-0": "^6.3.13",
+    "babel-register": "^6.4.3",
     "chai": "^3.4.1",
     "eslint": "^1.10.3",
     "lodash.omit": "^4.0.1",
