+++ test/.eslintrc.json
@@ -1,5 +1,8 @@
 {
   "env": {
     "mocha": true
+  },
+  "rules": {
+    "prefer-arrow-callback": [0]
   }
 }
