+++ test/mocha.opts
@@ -1,0 +1,3 @@
+--compilers js:babel-register
+--reporter spec
+--timeout 10000
